第三次上传，超声波测距没有问题，数值正确，最小测距距离为4cm

+++ src/Chao_Sheng_Bo.h
@@ -5,6 +5,6 @@
 #define TRIG_PIN      BSP_IO_PORT_04_PIN_13  // Trig引脚
 
 void Chao_Sheng_Bo_GPT0_Init(void); // 超声波定时器初始化
-uint32_t HC_SR04_Measure(void); // 超声波测距
+float	HC_SR04_Measure(void); // 超声波测距，最小检测距离4cm，100ms一检测最合理。
 
 #endif
